Track threads started from the room itself
Previously, we would create one off messages for pings coming directly
from the room, and only track threads in which pings were started from
pre-existing threads.

This is silly because once you ask a question to the bot from the
room, you should be able to continue that conversation within the
thread created when the bot replies, with the full context of the first
message.

Tweak things here so that we always create threads.

While we're here, make sure to only add the pre-prompt when creating new
chats, and not to every user message within a chat.